+++ Aurora/src/Aurora.h
@@ -6,6 +6,10 @@
 #include "Aurora/Layer.h"
 #include "Aurora/Log.h"
 
+#include "Aurora/Input.h"
+#include "Aurora/KeyCodes.h"
+#include "Aurora/MouseButtonCodes.h"
+
 #include "Aurora/ImGui/ImGuiLayer.h"
 
 // ===== Entry Point ===================================================