+++ Aurora/src/Aurora/Application.h
@@ -7,6 +7,8 @@
 #include "Events/Event.h"
 #include "Aurora/Events/ApplicationEvent.h"
 
+#include "Aurora/ImGui/ImGuiLayer.h"
+
 namespace Aurora {
 
 	class AURORA_API Application
@@ -29,6 +31,7 @@ namespace Aurora {
 		bool OnWindowClose(WindowCloseEvent& e);
 
 		std::unique_ptr<Window> m_Window;
+		ImGuiLayer* m_ImGuiLayer;
 		bool m_Running = true;
 		LayerStack m_LayerStack;
 	private: