+++ Aurora/src/Aurora/ImGui/ImGuiLayer.h
@@ -2,6 +2,10 @@
 
 #include "Aurora/Layer.h"
 
+#include "Aurora/Events/ApplicationEvent.h"
+#include "Aurora/Events/KeyEvent.h"
+#include "Aurora/Events/MouseEvent.h"
+
 namespace Aurora {
 
 	class AURORA_API ImGuiLayer : public Layer
@@ -10,10 +14,12 @@ namespace Aurora {
 		ImGuiLayer();
 		~ImGuiLayer();
 
-		void OnAttach();
-		void OnDetach();
-		void OnUpdate();
-		void OnEvent(Event& event);
+		virtual void OnAttach() override;
+		virtual void OnDetach() override;
+		virtual void OnImGuiRender() override;
+
+		void Begin();
+		void End();
 	private:
 		float m_Time = 0.0f;
 	};