+++ Sandbox/src/SandboxApp.cpp
@@ -1,5 +1,7 @@
 #include <Aurora.h>
 
+#include "imgui/imgui.h"
+
 class ExampleLayer : public Aurora::Layer {
 public:
 	ExampleLayer()
@@ -8,11 +10,25 @@ public:
 	}
 
 	void OnUpdate() override {
-		AU_INFO("ExampleLayer::Update()");
+		//if (Aurora::Input::IsKeyPressed(AU_KEY_TAB))
+		//	AU_TRACE("Tab key is pressed (poll)!");
+	}
+
+	virtual void OnImGuiRender() override
+	{
+		ImGui::Begin("Test");
+		ImGui::Text("Hello World");
+		ImGui::End();
 	}
 
 	void OnEvent(Aurora::Event& event) override {
-		AU_TRACE("{0}", event);
+		if (event.GetEventType() == Aurora::EventType::KeyPressed)
+		{
+			Aurora::KeyPressedEvent& e = (Aurora::KeyPressedEvent&)event;
+			if (e.GetKeyCode() == AU_KEY_TAB)
+				AU_TRACE("Tab key is pressed (event)!");
+			AU_TRACE("{0}", (char)e.GetKeyCode());
+		}
 	}
 };
 
@@ -20,7 +36,6 @@ class Sandbox : public Aurora::Application {
 public:
 	Sandbox() {
 		PushLayer(new ExampleLayer());
-		PushOverlay(new Aurora::ImGuiLayer());
 	}
 
 	~Sandbox() {