+++ premake5.lua
@@ -1,5 +1,6 @@
 workspace "Aurora"
 	architecture "x64"
+	startproject "Sandbox"
 
 	configurations
 	{
@@ -14,17 +15,20 @@ IncludeDir = {}
 IncludeDir["GLFW"] = "Aurora/vendor/GLFW/include"
 IncludeDir["Glad"] = "Aurora/vendor/Glad/include"
 IncludeDir["ImGui"] = "Aurora/vendor/imgui"
+IncludeDir["glm"] = "Aurora/vendor/glm"
 
-include "Aurora/vendor/GLFW"
-include "Aurora/vendor/Glad"
-include "Aurora/vendor/imgui"
+group "Dependencies"
+	include "Aurora/vendor/GLFW"
+	include "Aurora/vendor/Glad"
+	include "Aurora/vendor/imgui"
 
-startproject "Sandbox"
+group ""
 
 project "Aurora"
 	location "Aurora"
 	kind "SharedLib"
 	language "C++"
+	staticruntime "Off"
 	
 	targetdir ("bin/" .. outputdir .. "/%{prj.name}")
 	objdir ("bin-int/" .. outputdir .. "/%{prj.name}")
@@ -35,7 +39,9 @@ project "Aurora"
 	files
 	{
 		"%{prj.name}/src/**.h",
-		"%{prj.name}/src/**.cpp"
+		"%{prj.name}/src/**.cpp",
+		"%{prj.name}/vendor/glm/glm/**.hpp",
+		"%{prj.name}/vendor/glm/glm/**.inl"
 	}
 
 	includedirs
@@ -44,7 +50,8 @@ project "Aurora"
 		"%{prj.name}/vendor/spdlog/include",
 		"%{IncludeDir.GLFW}",
 		"%{IncludeDir.Glad}",
-		"%{IncludeDir.ImGui}"
+		"%{IncludeDir.ImGui}",
+		"%{IncludeDir.glm}"
 	}
 
 	links
@@ -69,28 +76,29 @@ project "Aurora"
 
 		postbuildcommands
 		{
-			("{COPY} %{cfg.buildtarget.relpath} ../bin/" .. outputdir .. "/Sandbox")
+			("{COPY} %{cfg.buildtarget.relpath} \"../bin/" .. outputdir .. "/Sandbox/\"")
 		}
 
 	filter "configurations:Debug"
 		defines "AU_DEBUG"
-		buildoptions "/MDd"
+		runtime "Debug"
 		symbols "On"
 		
 	filter "configurations:Release"
 		defines "AU_RELEASE"
-		buildoptions "/MD"
+		runtime "Release"
 		optimize "On"
 		
 	filter "configurations:Dist"
 		defines "AU_DIST"
-		buildoptions "/MD"
+		runtime "Release"
 		optimize "On"
 
 project "Sandbox"
 	location "Sandbox"
 	kind "ConsoleApp"
 	language "C++"
+	staticruntime "Off"
 	
 	targetdir ("bin/" .. outputdir .. "/%{prj.name}")
 	objdir ("bin-int/" .. outputdir .. "/%{prj.name}")
@@ -104,7 +112,9 @@ project "Sandbox"
 	includedirs
 	{
 		"Aurora/vendor/spdlog/include",
-		"Aurora/src"
+		"Aurora/src",
+		"Aurora/vendor",
+		"%{IncludeDir.glm}"
 	}
 
 	links
@@ -124,15 +134,15 @@ project "Sandbox"
 
 	filter "configurations:Debug"
 		defines "AU_DEBUG"
-		buildoptions "/MDd"
+		runtime "Debug"
 		symbols "On"
 		
 	filter "configurations:Release"
 		defines "AU_RELEASE"
-		buildoptions "/MD"
+		runtime "Release"
 		optimize "On"
 		
 	filter "configurations:Dist"
 		defines "AU_DIST"
-		buildoptions "/MD"
+		runtime "Release"
 		optimize "On"